# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Sheikh Enterprise Backend API
  description: API documentation for Sheikh Enterprise Backend
  version: 1.0.0
  contact:
    name: Raihanur Rahman
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "error message"
...
paths:
...
  /api/suppliers:
    get:
      tags:
        - Suppliers
      summary: List suppliers
      description: Get a paginated list of suppliers with optional filters
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: page_size
          schema:
            type: integer
            default: 10
          description: Page size
      responses:
        '200':
          description: Suppliers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Supplier'
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                      page_size:
                        type: integer
                      total:
                        type: integer

    post:
      tags:
        - Suppliers
      summary: Create supplier
      description: Create a new supplier
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: Supplier created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/suppliers/{id}:
    get:
      tags:
        - Suppliers
      summary: Get a supplier by ID
      description: Get detailed information about a supplier
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Supplier ID
      responses:
        '200':
          description: Supplier retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Suppliers
      summary: Update supplier
      description: Update an existing supplier
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Supplier ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '200':
          description: Supplier updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Suppliers
      summary: Delete supplier
      description: Mark a supplier as deleted
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Supplier ID
      responses:
        '200':
          description: Supplier deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "supplier deleted successfully"
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/shops:
    get:
      tags:
        - Shops
      summary: List shops
      description: Get a paginated list of shops with optional filters
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: page_size
          schema:
            type: integer
            default: 10
          description: Page size
        - in: query
          name: company_id
          schema:
            type: string
            format: uuid
          description: Filter by company ID
      responses:
        '200':
          description: Shops retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Shop'
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                      page_size:
                        type: integer
                      total:
                        type: integer

    post:
      tags:
        - Shops
      summary: Create shop
      description: Create a new shop
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShopRequest'
      responses:
        '201':
          description: Shop created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/shops/{id}:
    get:
      tags:
        - Shops
      summary: Get a shop by ID
      description: Get detailed information about a shop
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Shop ID
      responses:
        '200':
          description: Shop retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Shops
      summary: Update shop
      description: Update an existing shop
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Shop ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShopRequest'
      responses:
        '200':
          description: Shop updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Shops
      summary: Delete shop
      description: Mark a shop as deleted
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Shop ID
      responses:
        '200':
          description: Shop deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "shop deleted successfully"
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/companies/{company_id}/shops:
    get:
      tags:
        - Shops
      summary: Get shops by company ID
      description: Get all shops belonging to a specific company
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
            format: uuid
          description: Company ID
      responses:
        '200':
          description: Shops retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shop'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 